move ScaledPicassoImageView to base-tool module
