+++ ui/base-tool/src/main/java/org/cru/godtools/base/tool/picasso/transformation/ScaledCropTransformation.java
@@ -1,4 +1,4 @@
-package org.cru.godtools.tract.picasso.transformation;
+package org.cru.godtools.base.tool.picasso.transformation;
 
 import android.graphics.Bitmap;
 import android.support.annotation.NonNull;
