+++ ui/base-tool/src/main/res/values/attrs.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="ScaledPicassoImageView">
+        <attr name="scaleType" format="enum">
+            <enum name="fit" value="0" />
+            <enum name="fill" value="1" />
+            <enum name="fill_x" value="2" />
+            <enum name="fill_y" value="3" />
+        </attr>
+    </declare-styleable>
+</resources>
