+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/activity/TractActivity.java
@@ -41,6 +41,7 @@ import org.ccci.gto.android.common.util.os.BundleUtils;
 import org.cru.godtools.analytics.model.AnalyticsDeepLinkEvent;
 import org.cru.godtools.base.model.Event;
 import org.cru.godtools.base.tool.activity.ImmersiveActivity;
+import org.cru.godtools.base.tool.widget.ScaledPicassoImageView;
 import org.cru.godtools.download.manager.DownloadProgress;
 import org.cru.godtools.download.manager.GodToolsDownloadManager;
 import org.cru.godtools.model.Tool;
@@ -56,7 +57,6 @@ import org.cru.godtools.tract.content.TractManifestLoader;
 import org.cru.godtools.tract.util.DrawableUtils;
 import org.cru.godtools.tract.util.ViewUtils;
 import org.cru.godtools.tract.viewmodel.ManifestViewUtils;
-import org.cru.godtools.tract.widget.ScaledPicassoImageView;
 import org.cru.godtools.xml.model.Card;
 import org.cru.godtools.xml.model.Manifest;
 import org.cru.godtools.xml.model.Page;
