+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/ManifestViewUtils.java
@@ -5,8 +5,8 @@ import android.graphics.Typeface;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import org.cru.godtools.base.tool.widget.ScaledPicassoImageView;
 import org.cru.godtools.base.ui.util.LocaleTypefaceUtils;
-import org.cru.godtools.tract.widget.ScaledPicassoImageView;
 import org.cru.godtools.xml.model.Manifest;
 
 public class ManifestViewUtils {
