+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/PageViewHolder.java
@@ -15,9 +15,9 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
 
 import org.cru.godtools.base.model.Event;
+import org.cru.godtools.base.tool.widget.ScaledPicassoImageView;
 import org.cru.godtools.tract.R2;
 import org.cru.godtools.tract.widget.PageContentLayout;
-import org.cru.godtools.tract.widget.ScaledPicassoImageView;
 import org.cru.godtools.xml.model.Card;
 import org.cru.godtools.xml.model.Page;
