+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/ResourceViewUtils.java
@@ -9,7 +9,7 @@ import android.widget.ImageView;
 import android.widget.RelativeLayout;
 
 import org.ccci.gto.android.common.picasso.view.PicassoImageView;
-import org.cru.godtools.tract.widget.ScaledPicassoImageView;
+import org.cru.godtools.base.tool.widget.ScaledPicassoImageView;
 import org.cru.godtools.xml.model.ImageGravity;
 import org.cru.godtools.xml.model.ImageScaleType;
 import org.cru.godtools.xml.model.Resource;
