+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/widget/TractPicassoImageView.java
@@ -6,6 +6,7 @@ import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 
 import org.ccci.gto.android.common.picasso.view.SimplePicassoImageView;
+import org.cru.godtools.base.tool.widget.ScaledPicassoImageView;
 import org.cru.godtools.xml.model.ImageScaleType;
 
 import jp.wasabeef.picasso.transformations.CropTransformation.GravityHorizontal;
