+++ ui/tract-renderer/src/main/res/values/attrs.xml
@@ -10,13 +10,4 @@
         <attr name="layout_card_peek_toTopOf" format="reference" />
         <attr name="layout_card_stack_toTopOf" format="reference" />
     </declare-styleable>
-
-    <declare-styleable name="ScaledPicassoImageView">
-        <attr name="scaleType" format="enum">
-            <enum name="fit" value="0" />
-            <enum name="fill" value="1" />
-            <enum name="fill_x" value="2" />
-            <enum name="fill_y" value="3" />
-        </attr>
-    </declare-styleable>
 </resources>
